Todo.Application/Task/Queries/GetTaskById/GetTaskByIdQueryHandler.cs
using AutoMapper;
using MediatR;
using Todo.Application.Task.Dto;
using Todo.Domain.Exceptions;
using Todo.Domain.Repositories;

namespace Todo.Application.Task.Queries.GetTaskById;

public class GetTaskByIdQueryHandler : IRequestHandler<GetTaskByIdQuery, TaskDto>
{
    private readonly ITodoRepository _todoRepository;
    private readonly IMapper _mapper;
    public GetTaskByIdQueryHandler(ITodoRepository todoRepository, IMapper mapper)
    {
        _todoRepository = todoRepository;
        _mapper = mapper;
    }
    public async Task<TaskDto> Handle(GetTaskByIdQuery request, CancellationToken cancellationToken)
    {
        var task = await _todoRepository.GetById(request.TaskId);
        if (task is null) throw new NotFoundException("Task not found");
        var taskDto = _mapper.Map<TaskDto>(task);
        return taskDto;
    }
}
